Removed vendor direct dep; add libs, refer consts
