--- ext/src/options/js/vendors/caldom.min.mjs.js
@@ -1,1 +1,0 @@
-export default(function(){var t=window,n=Array.prototype,x=n.slice,i=t.Node.prototype,r=i.appendChild,o=i.insertBefore,y="isConnected"in Node.prototype?function(t){return t.isConnected}:function(t){return!(t.ownerDocument.compareDocumentPosition(t)&Node.DOCUMENT_POSITION_DISCONNECTED)},p=null!=t.Proxy,w=t.requestAnimationFrame||t.webkitRequestAnimationFrame||t.mozRequestAnimationFrame||function(t){return setTimeout(t,1e3/60)};t.Element.prototype.matches=t.Element.prototype.matches||t.Element.prototype.webkitMatchesSelector||t.Element.prototype.t||t.Element.prototype.msMatchesSelector||t.Element.prototype.oMatchesSelector;var m=function(t,i,n){this.init(t,i,n)};function j(t,i,n){if(t){var r,o=1==t.nodeType;if(!i)return o&&C(t),n.appendChild(t);var e=null!=t.i?(r=!0,t.i._h):t._h;if(o&&t.tagName==i.tagName&&e===i._h&&t.getAttribute("caldom-v")===i.getAttribute("caldom-v")){n=!0;if(!(t=r?t.i:t).isEqualNode(i)){e=t.hasChildNodes();if(e||i.hasChildNodes()){var u=i.childNodes,f=0;if(e)for(var n=!1,s=x.call(t.childNodes),h=s.length;f<h;f++){var c=s[f];if(j(c,u[f],i)&&1==c.nodeType){var a=x.call(c.querySelectorAll("* [_ref]"));a.push(c);for(var v=0,l=a.length;v<l;v++){var d=a[v],y=d._$k;y&&(d._$o[y]=d._$v)}}}for(var p=u.length;f<p;f++)i.lastChild._h?(b&&i.lastChild._h.o(),i.removeChild(i.lastChild),b&&i.lastChild._h.u()):i.removeChild(i.lastChild)}if(!t.isEqualNode(i)){for(var w=0,m=t.attributes.length;w<m;w++){var _=t.attributes[w];i.getAttribute(_.name)!=_.value&&i.setAttribute(_.name,_.value)}for(var g=0;g<i.attributes.length;g++){var A=i.attributes[g];t.hasAttribute(A.name)||(i.removeAttribute(A.name),g--)}}}!function t(i,n,r){if(Object.getPrototypeOf(n).hasOwnProperty("value")&&(n.value!=i.value&&(n.value=i.value),n.checked!=i.checked&&(n.checked=i.checked),n.indeterminate!=i.indeterminate&&(n.indeterminate=i.indeterminate)),n.selected!=i.selected&&(n.selected=i.selected),n._data!=i._data&&(n._data=i._data),r)for(var o=x.call(i.children),e=x.call(n.children),u=0,f=o.length;u<f;u++)t(o[u],e[u])}(t,i,n)}else{if(o){var b=null!=i._h;return b&&i._h.o(),o&&C(t),i.parentNode.replaceChild(r?t.i:t,i),b&&i._h.u(),t}i.nodeValue=t.nodeValue}}}function C(t){for(var i=x.call(t.querySelectorAll("* [caldom]")),n=0,r=i.length;n<r;n++){var o=i[n],e=o.i.parentNode,u=o.i.nextSibling;o.parentNode.replaceChild(o.i,o),e.insertBefore(o,u)}}function e(){this.each=h,this.map=b,this.find=P,this.prop=c,this.attr=l,this.css=v,this.addClass=_,this.removeClass=A}function u(t,i,n){i instanceof m&&(i=i.elems[0]);for(var r=0,o=t.length;r<o;r++){var e=t[r];if(e){var u=!!e.elems,f=u?(e.h(),e.v||e.elems):a(e)?e:[e];void 0===i&&(i=this.elem.firstChild);for(var s=0,h=f.length;s<h;s++){var c=f[s];c&&("object"!=typeof c&&(c=this._w.document.createTextNode(c)),n.call(this.elem,c,i))}u&&(e.v=null,e.l())}}return this}function f(t,i){i=i||window.document;var n=(t=t.trim())[0];if("+"==n)return[i.createElement(t.substr(1))];if("$"!=n)return"<"==n?x.call(i.createRange().createContextualFragment(t).childNodes):"~"==n?[i.createElementNS("http://www.w3.org/2000/svg",t.substr(1))]:x.call(i.querySelectorAll(t));for(var r=(i.ownerDocument||i).evaluate(t.substr(1),i,null,0,null),o=[],e=r.iterateNext();e;)o.push(e),e=r.iterateNext();return o}function s(t){return Array.isArray(t)?t:t.split(" ")}function a(t){return Array.isArray(t)||t&&"string"!=typeof t&&null!=t.length&&!(t instanceof Node)}function h(t){for(var i=0,n=this.elems.length;i<n&&!1!==t.call(this.elems[i],this.elems[i],i);i++);return this}function c(n,r,t){if(t||"string"==typeof n){if(null==r)return this.map(function(t){return t[n]});Array.isArray(r)?this.each(function(t,i){t[n]=r[i]}):this.each(function(t){t[n]=r})}else this.each(function(t){for(var i in n)t[i]=n[i]});return this}function v(n,r){if("string"!=typeof n)this.each(function(t){for(var i in n)t.style[i]=n[i]});else{if(null==r)return this.map(function(t){return t.style[n]});Array.isArray(r)?this.each(function(t,i){t.style[n]=r[i]}):this.each(function(t){t.style[n]=r})}return this}function l(n,r){if("string"!=typeof n)this.each(function(t){for(var i in n)t.setAttribute(i,n[i])});else{if(2!=arguments.length)return this.map(function(t){return t.getAttribute(n)});Array.isArray(r)?this.each(function(t,i){t.setAttribute(n,r[i])}):this.each(function(t){t.setAttribute(n,r)})}return this}function d(t,i){for(var n=0,r=i.length;n<r;n++)t.classList.add(i[n])}function _(t){var i;return t&&(i=s(t),this.each(function(t){d(t,i)})),this}function g(t,i){for(var n=0,r=i.length;n<r;n++)t.classList.remove(i[n])}function A(t){var i;return t&&(i=s(t),this.each(function(t){g(t,i)})),this}function b(t){for(var i=[],n=0,r=this.elems.length;n<r;n++)i.push(t(this.elems[n],n));return i}function P(i){var t=this.map(function(t){return x.call(f(i,t))});return new m(n.concat.apply([],t),void 0,this._w)}m.prototype={init:function(t,i,n){return this._w=n||window,this.$={},t&&(a(t)?this.elems=x.call(t):this.elems="object"==typeof t?[t]:f(t,this._w.document),this.elem=this.elems[0],1!=this.elems.length&&e.call(this)),i&&this.append(i),this},elems:[],elem:void 0,find:function(t){return new m(f(t,this.elem),void 0,this._w)},eq:function(t){return new m(this.elems[t<0?this.elems.length+t:t],void 0,this._w)},parent:function(r){var o=this,e=[];return r&&"number"!=typeof r?this.each(function(t){for(var i=t.parentNode,n=i==o._w.document;!n&&!i.matches(r);)n=(i=i.parentNode)==o._w.document;n||e.push(i)}):this.each(function(t){for(var i=t.parentNode,n=0;n<r;n++)i=i.parentNode;e.push(i)}),new m(e,void 0,this._w)},children:function(e){var u=[];return this.each(function(t){for(var i=x.call(t.childNodes),n=0,r=i.length;n<r;n++){var o=i[n];(!e||o.matches&&o.matches(e))&&u.push(o)}}),new m(u,void 0,this._w)},each:function(t){t.call(this.elem,this.elem,0)},map:function(t){return[t(this.elem,0)]},html:function(t){return this.prop("innerHTML",t,!0)},text:function(t){return this.prop("textContent",t,!0)},val:function(t){return this.prop("value",t,!0)},attr:function(t,i){if("string"!=typeof t)for(var n in t)this.elem.setAttribute(n,t[n]);else{if(2!=arguments.length)return[this.elem.getAttribute(t)];Array.isArray(i)?this.elem.setAttribute(t,i[0]):this.elem.setAttribute(t,i)}return this},prop:function(t,i,n){if(n||"string"==typeof t){if(null==i)return[this.elem[t]];Array.isArray(i)?this.elem[t]=i[0]:this.elem[t]=i}else for(var r in t)this.elem[r]=t[r];return this},data:function(n,r){return null==r?this.map(function(t){if(t._data)return t._data[n]}):(Array.isArray(r)?this.each(function(t,i){t._data||(t._data={}),t._data[n]=r[i]}):this.each(function(t){t._data||(t._data={}),t._data[n]=r}),this)},css:function(t,i){if("string"!=typeof t)for(var n in t)this.elem.style[n]=t[n];else{if(null==i)return[this.elem.style[t]];Array.isArray(i)?this.elem.style[t]=i[0]:this.elem.style[t]=i}return this},addClass:function(t){t&&d(this.elem,s(t));return this},removeClass:function(t){t&&g(this.elem,s(t));return this},show:function(t){return this.css("display",t||"block")},hide:function(){return this.css("display","none")},on:function(t,r,o){var e=s(t);return this.each(function(t){for(var i=0,n=e.length;i<n;i++)t.addEventListener(e[i],r,o)}),this},off:function(t,r,o){var e=s(t);return this.each(function(t){for(var i=0,n=e.length;i<n;i++)t.removeEventListener(e[i],r,o)}),this},append:function(t){return u.call(this,a(t)?t:arguments,null,r)},prepend:function(t,i){return u.call(this,a(t)?t:[t],i,o)},remove:function(){return this.o(!0),this.each(function(t){t.parentNode.removeChild(t)}),this.elems=[],this.elem=void 0,e.call(this),this.u(!0),this},replace:function(n){return this.each(function(t,i){i=n(t,i,this);j(i=i instanceof m?i.elems[0]:i,t,t)}),this},h:function(){(this.render||this.update)&&(this.willMount&&!this.p&&this.willMount(this),this.react(void 0,void 0,void 0,void 0,void 0,!0))},l:function(){var t=1==this.p;this.p=!0,this.didMount&&!t&&this.didMount(this)},o:function(t){this.willUnmount&&this.willUnmount(this,t)},u:function(t){this.p=!1,this.didUnmount&&this.didUnmount(this,t)},m:function(){this._={},this.g=0,this.didUpdate&&this.didUpdate(this)},react:function(t,i,n,r,o,e,u){var f,s=this,h=0==arguments.length;if(s.watched&&h)return s;if(n||h&&!u)return s.update&&(h||s.g++,s._[r]=o),s.A?void 0:(s.A=w(function(){s.react(void 0,void 0,void 0,void 0,void 0,void 0,!0)}),s);if(!e){if(i)for(var c in i)s[c]=i[c];p||(s.watched=!1),null!=t&&(0!=s.watched?s.state=(f=function(t,i){s.react(void 0,void 0,!0,t,i)},new Proxy(t,{get:function(t,i){t=t[i];return"object"!=typeof t||null==t||t instanceof m||"_"==i[0]?t:new Proxy(t,this)},set:function(t,i,n){return t[i]=n,f(i,!0),!0},deleteProperty:function(t,i){return delete t[i],f(i,!1),!0}})):s.state=t)}s.A=null;var a=s.elem,v=!!a;if(!s.p&&!e){if(!v||!y(a))return s;s.p=!0}h=s.p&&e;if(!h&&s.update&&(!s.render||v&&a._h)&&!s.update(s.state,s,s._,s.g))return s.m(),s;if(h){var l=a.cloneNode(!0);l._h=s,l.setAttribute("caldom",""),l.i=a,s.v=[l]}else{l=s.render(s.state,s);if(!l)return s;if((l=l instanceof m?l.elems[0]:l)._h=s,l.setAttribute("caldom",""),a){a=j(l,a,a);a&&(s.elem=a)}else{function d(t,i,n){var r=n instanceof m?n.elems[0]:n;r instanceof Element&&!r._$k&&(r._$o=t,r._$k=i,r._$v=n,r.setAttribute("_ref",i))}for(var c in s.elems=[l],this.elem=this.elems[0],s.$)d(s.$,c,s.$[c]);p&&(s.$=new Proxy(s.$,{set:function(t,i,n){return d(t,i,n),t[i]||(t[i]=n),!0}}))}s.m()}return s},watched:void 0,$:void 0,_w:void 0,_:{},g:0};function M(t,i,n){return new m(t,i,n)}return M.q=f,M.Component=m,M.version="1.0.7","undefined"!=typeof module&&module.exports?module.exports=M:"function"==typeof define?define(function(){return M}):t[t._caldom_alias||"_"]=M,M})();
