+++ ext/src/options/options.html
@@ -366,9 +366,6 @@
     <button id="resore-defaults">Restore Defaults</button>
     <div id="status"></div>
   </section>
-  <script src="./js/vendors/petite-vue.iife.js"></script>
-  <script type="module" src="./js/components/CRCheckbox.js"></script>
-  <script type="module" src="./js/quickDubsOptions.js"></script>
   <script type="module" src="./js/options.js"></script>
 </body>
